requirements.txt
scikit-learn==0.24.1
numpy==1.23.5
nltk==3.8.1
ufal.udpipe==1.3.0.1